Regression in pattern matching optional fields.
Fixes https://github.com/rescript-lang/rescript/issues/7400

Fix regression in pattern  matching for optional fields.

There's an optimisation where `{field: VariantCase}` is compiled to `field:VariantCase` instead of `field:Some(VariantCase)` to avoid generating code that handles undefined, since `VariantCase` is going to be a string.
However, that optimisation happens early in the pattern matching compiler, and can end up interfering with other cases. For example, when `VariantCase` is the only case in a variant, and the following case is `_`, the compiler is induced in thinking that the variant case covers all cases, and the case for `_` is never generated.
Similar behaviour happens when handling the last case of a variant followed by `_`.