Fix missing checks for duplicate literals in variants with payloads.
Fixes https://github.com/rescript-lang/rescript/issues/7438

In variant definitions without payloads, an error is reported if two cases have the same literal `@as(literal)`.
The check so far was missing for cases with payload, where the literal goes into the tag.
However it's perfectly possible to use the same tag in a case without payload, and one with payloads, as the two representations don't overlap.